deleteUserFile check folder contains file and return Status

+++ src/main/java/service/UserFileService.java
@@ -40,10 +40,13 @@ public class UserFileService {
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     @Path("/deleteFile")
-    public boolean deleteUserFile(Token token, @Context HttpServletRequest request) throws IOException {
+    public Status deleteUserFile(Token token, @Context HttpServletRequest request) throws IOException {
         LOG.info("deleteUserFile method is called with token:{}, id: {}", token.getToken(), token.getId());
         int fileId = getFileId(token, "deleteUserFile");
         UserFile userFile = userFileController.getUserFile(fileId);
+        if(userFile.getSize() > 0 && userFile.isFolder()) {
+            return new Status(Operation.USERFILE, false, "This folder is not empty!");
+        }
         UserFileManager.deleteFile(userFile.getPath() + "\\" + userFile.getId() + userFile.getExtension());
         double fileSize = -userFile.getSize();
         int parentId = userFile.getParentId();
@@ -52,7 +55,10 @@ public class UserFileService {
         if (folderParentId != 1) {
             userFileController.changeFolderCurrSize(folderParentId, -fileSize);
         }
-        return userFileController.deleteUserFile(token.getId());
+        if (userFileController.deleteUserFile(token.getId())) {
+            return new Status(Operation.USERFILE, true, "Folder/File successfully deleted!");
+        }
+        return new Status(Operation.USERFILE, false, "There was an error during deleting!");
     }
 
     @POST
